requirements.txt
absl-py==0.9.0
asn1crypto==1.3.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
future==0.18.2
glob2==0.7
google-auth==1.15.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
hydra-core==0.11.3
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.3
jupyterlab-server==1.1.5
kiwisolver==1.2.0
libarchive-c==2.9
lief==0.9.0
line-profiler==3.0.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.3.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
omegaconf==1.4.1
opencv-python==4.2.0.34
packaging==20.4
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.4
protobuf==3.12.1
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.2
python-dateutil==2.8.1
pytorch-lightning==0.7.6
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel-yaml==0.15.87
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
torch==1.5.0
torch-summary==1.3.1
torchvision==0.6.0a0+82fd1c8
tornado==6.0.4
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
